Orders route started

+++ front-end/src/components/Userproduct/Viewproduct.jsx
@@ -1,38 +1,42 @@
 import React from "react";
-import { useState, useEffect, CSSProperties } from "react";
+import { useState, useEffect } from "react";
 import axios from "axios";
-import "./Userviewproduct.css";
+import "./Viewproduct.css";
 import { Link, useNavigate } from "react-router-dom";
-import Navbar from "../Navbar/Navbar";
 import toast, { Toaster } from "react-hot-toast";
 // import Usernavbar from "../UserNavbar/Usernavbar";
-const Userviewproduct = () => {
+const Viewproduct = () => {
   const navigate = useNavigate();
   const [product, setproduct] = useState([]);
   const token = localStorage.getItem("Token");
+  const role = localStorage.getItem("Role");
   useEffect(() => {
-    axios
-      .get(`http://localhost:2222/api/user/view`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      })
-      .then((data) => {
-        console.log(data.data.data);
-        setproduct(data.data.data);
-      })
-      .catch((err) => {
-        console.log(err);
-        if (err.response.status == 401) {
-          toast.error("Session Time Out",{
-            position:'bottom-center'
-          });
-          setTimeout(() => {
-            localStorage.clear();
-            navigate("/login");
-          }, 3000);
-        }
-      });
+    if (role == 2) {
+      axios
+        .get(`http://localhost:2222/api/user/view`, {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        })
+        .then((data) => {
+          console.log(data.data.data);
+          setproduct(data.data.data);
+        })
+        .catch((err) => {
+          console.log(err);
+          if (err.response.status == 401) {
+            toast.error("Session Time Out", {
+              position: "bottom-center",
+            });
+            setTimeout(() => {
+              localStorage.clear();
+              navigate("/login");
+            }, 3000);
+          }
+        });
+    } else {
+      // sdf
+    }
   }, []);
 
   const deleteHandler = (id) => {
@@ -57,13 +61,12 @@ const Userviewproduct = () => {
       })
       .then((data) => {
         console.log(data);
-        toast.success("Add to cart successfully !",  {
+        toast.success("Add to cart successfully !", {
           position: "bottom-center",
-          
         });
         setTimeout(() => {
           navigate("/usercart");
-        }, 2000);
+        }, 1000);
       })
       .catch((err) => {
         console.log(err);
@@ -81,7 +84,7 @@ const Userviewproduct = () => {
       {product.length > 0 ? (
         <div className="user-view-content-body">
           <div className="user-view-cards-body">
-            <div className="user-view-cards-heading">PRODUCTS LISTS</div>
+            <div className="user-view-cards-heading">PRODUCTS LISTS </div>
             {product.map((item) => (
               <div className="user-view-card" key={item._id}>
                 <div className="user-view-card-image">
@@ -93,10 +96,22 @@ const Userviewproduct = () => {
                 </div>
                 <div className="user-view-card-details">
                   <h3 className="user-view-card-details-h3"> {item.name}</h3>
-                  <h4 className="user-view-card-details-h4"> {item.available_qty}</h4>
-                  <h4 className="user-view-card-details-h4"> {item.category}</h4>
-                  <h4 className="user-view-card-details-h4"> {item.sub_category}</h4>
-                  <h4 className="user-view-card-details-h4"> {item.description}</h4>
+                  <h4 className="user-view-card-details-h4">
+                    {" "}
+                    {item.available_qty}
+                  </h4>
+                  <h4 className="user-view-card-details-h4">
+                    {" "}
+                    {item.category}
+                  </h4>
+                  <h4 className="user-view-card-details-h4">
+                    {" "}
+                    {item.sub_category}
+                  </h4>
+                  <h4 className="user-view-card-details-h4">
+                    {" "}
+                    {item.description}
+                  </h4>
                   {/* {anime == true ? (
                     <img
                       src="/addtocartanimation.gif"
@@ -141,4 +156,4 @@ const Userviewproduct = () => {
   );
 };
 
-export default Userviewproduct;
+export default Viewproduct;
